refactor: move from `wrap` to `iife`.
