+++ crates/rolldown/src/ecmascript/format/utils/mod.rs
@@ -1,3 +1,2 @@
 pub(super) mod external_module;
 pub(super) mod namespace;
-pub mod wrapper;
